committing Building Project

+++ Q4_Partheev_building/src/main/java/com/building/Q4_Partheev_building/BuildingEntity.java
@@ -12,17 +12,17 @@ public class BuildingEntity {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	private Integer floor_num;
-	private Float floor_area;
+	private Double floor_area;
 	public Integer getFloor_num() {
 		return floor_num;
 	}
 	public void setFloor_num(Integer floor_num) {
 		this.floor_num = floor_num;
 	}
-	public Float getFloor_area() {
+	public Double getFloor_area() {
 		return floor_area;
 	}
-	public void setFloor_area(Float floor_area) {
+	public void setFloor_area(Double floor_area) {
 		this.floor_area = floor_area;
 	}	
 }
